add: joblib backend, postprocessing-safe mutations, basic logger, serialization, mp4 visualization,  GA & SEPA2 on new core. Now fitness evaluates only for changed structures.

+++ gefest/core/geometry/domain.py
@@ -18,7 +18,7 @@ class Domain:
     max_poly_num: int = 4
     min_points_num: int = 20
     max_points_num: int = 50
-    polygon_side: int = 0.05
+    polygon_side: float = 0.05
     min_dist_from_boundary: float = 1.0
     prohibited_area: Optional[Structure] = Field(default=Structure([]))
     fixed_points: Optional[Union[Polygon, list[list[float]]]] = Field(default_factory=list)
